feat(499): RigProject.midi — project-level MIDI bindings
ADR 0003 says MIDI bindings belong to the project so they travel with the
.openrig file: the same setlist plays identically on every machine. Add a
`midi: Option<RigProjectMidi>` field to `RigProject` with a single
`bindings: Vec<Binding>` member. `None` for pre-#499 projects — they
fall back to the system bindings file or the shipped default at resolve
time (resolver lands in a later commit).

- Round-trip preserved: pre-#499 YAML parses with midi=None; `midi` field
  is skipped on serialize when None, so projects without bindings emit no
  phantom `midi: null` line.
- RED-first: `crates/project/src/rig_midi_tests.rs` (4 tests) failed to
  compile before the field existed — GREEN after.

Required refactor — `rig.rs` was already 624 lines (over the 600-line
hard cap) before this change. Adding the field would push it further out
of compliance. Extracted `impl RigProject` (the bulk: write-back,
preset/scene CRUD, validation) into `crates/project/src/rig_methods.rs`.
Pure move, no behavior change.

- `rig.rs` now 235 lines (struct/enum/preset-impl definitions only).
- `rig_methods.rs` new, 407 lines (impl RigProject block).
- Both under the 600-line cap.

All 9 test/migrate `RigProject { … }` literals updated to include
`midi: None,`. Workspace tests: 37 binaries, 0 failures.

Issue: #499
