+++ crates/adapter-gui/src/chain_rig_nav_repro_tests.rs
@@ -39,6 +39,7 @@ fn rig() -> RigProject {
         inputs,
         outputs: BTreeMap::new(),
         presets,
+        midi: None,
     }
 }
