+++ crates/adapter-gui/src/chain_rig_nav_tests.rs
@@ -45,6 +45,7 @@ fn rig() -> RigProject {
         inputs,
         outputs: BTreeMap::new(),
         presets,
+        midi: None,
     }
 }
