+++ crates/application/src/local_dispatcher_rig_tests.rs
@@ -61,6 +61,7 @@ fn rig() -> RigProject {
         inputs,
         outputs: BTreeMap::new(),
         presets,
+        midi: None,
     }
 }
