+++ crates/project/src/lib.rs
@@ -8,6 +8,7 @@ pub mod param;
 pub mod project;
 pub mod rig;
 pub mod rig_command;
+pub mod rig_methods;
 pub mod rig_sync;
 pub mod vst3_editor;
