+++ crates/project/src/rig_scene_tests.rs
@@ -51,6 +51,7 @@ fn project_with(inputs: Vec<(&str, RigInput)>, presets: &[&str]) -> RigProject {
                 )
             })
             .collect(),
+        midi: None,
     }
 }
