+++ crates/project/src/rig_tests.rs
@@ -72,6 +72,7 @@ fn project_with(inputs: Vec<(&str, RigInput)>, presets: &[&str]) -> RigProject {
                 )
             })
             .collect(),
+        midi: None,
     }
 }
